Create print_date_time.py
This script has a function that outputs different dates and formats functions, but this function needs to pass in a string as a parameter. If you do not want to pass in parameters, please pass in an empty string. There are many formats in the script annotation. Please read the script carefully.